Solving build problems, solving table creation problems, creating JPA repositories, setting application configs

+++ src/main/java/com/market/application/repositories/UserRepository.java
@@ -3,7 +3,7 @@ package com.market.application.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.market.entities.User;
-
-public interface UserRepository extends JpaRepository<User, Long>{
+	
+public interface UserRepository extends JpaRepository<User, Integer> {
 
 }
